Reorganize test project structure
